aimonk_multilabel/requirements.txt
torch
torchvision
matplotlib
Pillow
numpy
